% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_input.R
\name{complete_input}
\alias{complete_input}
\title{Fill in All Missing Values}
\usage{
complete_input(
  .data,
  waterfall = list(get_duty, get_type, get_model_year, get_purchase_year, get_annual_vmt,
    get_term)
)
}
\description{
Fill in All Missing Values
}
