% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_term_calc.R
\name{full_term}
\alias{full_term}
\title{Calculate GHG emissions and TCO of a vehicle(s) over its term}
\usage{
full_term(.data)
}
\arguments{
\item{.data}{vehicle input data, one row per vehicle}
}
\description{
Calculate GHG emissions and TCO of a vehicle(s) over its term
}
