% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_input.R
\name{get_duty}
\alias{get_duty}
\title{Filling in value for missing duty}
\usage{
get_duty(.data)
}
\arguments{
\item{.data}{Dataframe of vehicle data}
}
\description{
Fills in a value for duty for vehicle's missing it. Assumes
light duty.
}
\examples{
\dontrun{
get_duty(vehicle_data)
}
}
