% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_input.R
\name{get_initial_cost}
\alias{get_initial_cost}
\title{Fill in Missing Cost}
\usage{
get_initial_cost(.data)
}
\arguments{
\item{.data}{Dataframe of vehicle data}
}
\description{
Fills in value for vehicles missing purchase / leasing price,
uses the internal \code{initial_cost_df}.
}
\examples{
\dontrun{
get_initial_cost(vehicle_data)
}
}
