% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_input.R
\name{get_model_year}
\alias{get_model_year}
\title{Fill in Model Year}
\usage{
get_model_year(.data)
}
\arguments{
\item{.data}{Dataframe of vehicle data}
}
\description{
Adds a value for moddel year for vehicles missing this value,
assigns current year as the value
}
\examples{
\dontrun{
get_model_year(vehicle_data)
}
}
