% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_input.R
\name{get_type}
\alias{get_type}
\title{Filling in Value for Missing Type}
\usage{
get_type(.data)
}
\arguments{
\item{.data}{Dataframe of vehicle data}
}
\description{
Assigns a value for purchase type for vehicles missing a value,
assumes a default of "PURCHASE"
}
\examples{
\dontrun{
get_type(vehicle_data)
}
}
